+++ packages/pn-pa-webapp/src/utility/validation.utility.ts
@@ -7,6 +7,7 @@ import {
   NewNotificationF24Payment,
   NewNotificationPagoPaPayment,
   NewNotificationRecipient,
+  NotificationFeePolicy,
   RecipientPaymentsFormValues,
 } from '../models/NewNotification';
 import { getDuplicateValuesByKeys } from './notification.utility';
@@ -296,4 +297,26 @@ export function identicalSHA(values: RecipientPaymentsFormValues | undefined): A
   }
 
   return errors;
+}
+
+export function paFeeValidationSchema(tc: any) {
+  return yup
+    .string()
+    .optional()
+    .when('notificationFeePolicy', {
+      is: NotificationFeePolicy.DELIVERY_MODE,
+      then: yup.string().required(tc('required-field')),
+    });
+}
+
+export function isCurrencyAndMaxValue(
+  paFee: string | undefined
+): { messageKey: string; data?: { [key: string]: number | string } } | undefined {
+  if (paFee && !dataRegex.currency.test(String(paFee))) {
+    return { messageKey: 'notification-fee.pa-fee-currency' };
+  }
+  if (paFee && !dataRegex.maxOneEuro.test(String(paFee))) {
+    return { messageKey: 'notification-fee.pa-fee-max-value', data: { maxValue: 1 } };
+  }
+  return undefined;
 }